feat: use HCS metrics in util in all of the init scripts
- also add new scripts in init to assist in creating and tracking HCS metrics

+++ init/00-main.sh
@@ -4,8 +4,8 @@ DIR=$( cd -- "$( dirname -- "${BASH_SOURCE[0]}" )" &> /dev/null && pwd )
 
 cd ${DIR}/..
 npm install
-cp .env.sample .env
-cp .rpcrelay.env.sample .rpcrelay.env
+[ ! -f .env ] && .env.sample .env
+[ ! -f .rpcrelay.env ] && cp .rpcrelay.env.sample .rpcrelay.env
 
 cd ${DIR}
